+++ src/main/java/team01/yaksutor/admin/mapper/AdminMapper.java
@@ -1,0 +1,8 @@
+package team01.yaksutor.admin.mapper;
+
+public interface AdminMapper {
+
+    public int medicineInsert();
+
+    public String autoIncrementMediCode();
+}
