--- src/main/java/team01/yaksutor/admin/medicine/dto/Efficacy.java
@@ -1,12 +1,0 @@
-package team01.yaksutor.admin.medicine.dto;
-
-import lombok.Data;
-
-@Data
-public class Efficacy {
-    private String efficacyCode;
-    private String mediCode;
-    private String regMId;
-    private String efficacy;
-    private String regDate;
-}
