--- src/main/java/team01/yaksutor/admin/member/dto/member.java
@@ -1,4 +1,0 @@
-package team01.yaksutor.admin.member.dto;
-
-public class member {
-}
