--- src/main/java/team01/yaksutor/admin/member/mapper/MemberMapper.java
@@ -1,4 +1,0 @@
-package team01.yaksutor.admin.member.mapper;
-
-public interface MemberMapper {
-}
