--- src/main/java/team01/yaksutor/admin/member/service/MemberService.java
@@ -1,4 +1,0 @@
-package team01.yaksutor.admin.member.service;
-
-public interface MemberService {
-}
