+++ src/main/java/team01/yaksutor/dto/Member.java
@@ -1,0 +1,21 @@
+package team01.yaksutor.dto;
+
+import lombok.Data;
+
+@Data
+public class Member {
+    private String memberId;
+    private int memberLevelNum;
+    private String memberPw;
+    private String memberName;
+    private String memberGender;
+    private String memberEmail;
+    private String memberBirth;
+    private String memberPhone;
+    private String memberAddrSido;
+    private String memberAddrSigungu;
+    private String memberAddrDong;
+    private String memberPostCode;
+    private String memberAddr;
+
+}
