+++ src/main/java/team01/yaksutor/dto/MemberLevel.java
@@ -1,0 +1,10 @@
+package team01.yaksutor.dto;
+
+import lombok.Data;
+
+@Data
+public class MemberLevel {
+    private int levelNum;
+    private String levelName;
+    private String levelDate;
+}
