+++ src/main/java/team01/yaksutor/pharmacy/mapper/PhMemberMapper.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.pharmacy.mapper;
+
+public class PhMemberMapper {
+}
