--- src/main/java/team01/yaksutor/pharmacy/member/dto/MemberLevel.java
@@ -1,10 +1,0 @@
-package team01.yaksutor.pharmacy.member.dto;
-
-import lombok.Data;
-
-@Data
-public class MemberLevel {
-    private int levelNum;
-    private String levelName;
-    private String levelDate;
-}
