--- src/main/java/team01/yaksutor/pharmacy/member/mapper/MemberMapper.java
@@ -1,4 +1,0 @@
-package team01.yaksutor.pharmacy.member.mapper;
-
-public class MemberMapper {
-}
