+++ src/main/java/team01/yaksutor/supplier/mapper/SuMapper.java
@@ -1,0 +1,4 @@
+package team01.yaksutor.supplier.mapper;
+
+public interface SuMapper {
+}
